+++ spec/env/Page1.js
@@ -7,6 +7,10 @@ export class Page1 extends Page {
 		return 'Page1';
 	}
 
+	getTitle() {
+		return Promise.resolve(this.getClassName());
+	}
+
 	$render() {
 		return <span>Page1</span>;
 	}