+++ src/Page.js
@@ -9,12 +9,24 @@ export class Page extends Component {
 		this._node = null;
 	}
 
+	componentWillMount() {
+		var title = this.getTitle().then((title) => {
+			if (title && title.toString()) {
+				this.props.router.setTitle(title.toString());
+			}
+		});
+	}
+
 	$render() {
 		throw new Error('Page.$render is abstract');
 	}
 
 	getNode() {
 		return this._node;
+	}
+
+	getTitle() {
+		return Promise.resolve("adsfa");
 	}
 
 	render() {